api_yamdb/requirements.txt
APScheduler==3.6.3
asgiref==3.2.10
atomicwrites==1.4.1
attrs==22.1.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==2.2.16
django-filter==21.1
django-phone-field==1.8.1
django-phonenumber-field==6.3.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==3.9.2
flake8-docstrings==1.6.0
gunicorn==20.0.4
idna==3.3
importlib-metadata==1.7.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.54
Pillow==7.0.0
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.5
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytils==0.3
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
toml==0.10.2
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.5
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
